pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jam-model-memory"
version = "0.0.1"
description = "JAM (Journalistic Agent Memory) - Local-first memory system for LLM agents"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "JAM Team"},
]
maintainers = [
    {name = "JAM Team"},
]
requires-python = ">=3.11"
keywords = ["llm", "memory", "agent", "5w1h", "journalistic", "faiss", "semantic-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core Framework
    "fastapi>=0.116.1",
    "flask>=3.0.0",
    "jinja2>=3.1.4",
    "pydantic>=2.6.1",
    "structlog>=24.1.0",
    "uvicorn>=0.35.0",

    # Scientific Computing
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "umap-learn>=0.5.9",

    # Deep Learning
    "networkx>=3.0",
    "torch>=2.0.0",

    # Vector Search & Text Processing
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "tiktoken>=0.5.0",

    # LLM Integration
    "llama-cpp-python>=0.2.76",

    # Web & API
    "aiohttp>=3.12.15",
    "click>=8.1.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",

    # Utilities
    "psutil>=5.9.5",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov",
    "pytest-asyncio>=0.21.1",
]
docs = [
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
    "Pillow>=10.0.0",
    "PyPDF2>=3.0.0",
    "pytesseract>=0.3.10",
    "python-docx>=1.0.0",
    "PyYAML>=6.0",
    "toml>=0.10.2",
    "trafilatura>=2.0.0",
]
search = [
    "duckduckgo-search>=5.0.0",
]
viz = [
    "matplotlib>=3.10.6",
]
all = [
    "jam-model-memory[dev,docs,search,viz]",
]

[project.urls]
Homepage = "https://github.com/yourusername/jam-model-memory"
Documentation = "https://github.com/yourusername/jam-model-memory#readme"
Repository = "https://github.com/yourusername/jam-model-memory.git"
Issues = "https://github.com/yourusername/jam-model-memory/issues"

[project.scripts]
jam = "agentic_memory.cli:main"
jam-server = "agentic_memory.server.flask_app:main"
jam-api = "llama_api:main"
jam-analyze = "retrieval_analyzer:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_memory*", "data*"]
exclude = ["tests*", "logs*", "*.egg-info", "__pycache__"]

[tool.setuptools.package-data]
"agentic_memory.server" = ["templates/*.html", "static/css/*.css", "static/js/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false